Move HeatMapBoard to own file
