+++ src/logic/calculateHeatMap.ts
